Display

+++ app/src/test/java/com/example/socialcompass/MockOrientationTest.java
@@ -8,10 +8,10 @@ import static org.junit.Assert.*;
 
 @RunWith(RobolectricTestRunner.class)
 public class MockOrientationTest {
-    private static float delta = 0.001f;
+    private static double delta = 0.001f;
     @Test
     public void testNewOrientation() {
-        float orientation1 = 1.0f;
+        double orientation1 = 1.0;
         MockOrientation mo = new MockOrientation(orientation1);
         assertEquals(orientation1, mo.getOrientation(), delta);
     }
@@ -19,12 +19,12 @@ public class MockOrientationTest {
     @Test
     public void testDefaultOrientation() {
         MockOrientation mo = new MockOrientation();
-        assertEquals(0.0f, mo.getOrientation(), delta);
+        assertEquals(0.0, mo.getOrientation(), delta);
     }
 
     @Test
     public void testSetOrientation() {
-        float orientation2 = 1.5f;
+        double orientation2 = 1.5;
         MockOrientation mo = new MockOrientation();
         mo.setOrientation(orientation2);
         assertEquals(orientation2, mo.getOrientation(), delta);
